HELEN: READ COMMENT
I made changes that I think will fix the problem.

I reset the encoders as if we were starting the program at max speed, and decreased it the same way we do in the acceleration block.

Please take a look at this and make any changes, either to this version, or the version prior.